# Phase-II-III-IV/specs/001-phase2-todo-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Phase II Todo API
  version: 1.0.0
  description: REST API for Todo App with JWT Authentication
servers:
  - url: http://localhost:8000
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 1000
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        is_completed:
          type: boolean
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        is_completed:
          type: boolean
        owner_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
security:
  - BearerAuth: []
paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Requesting other users data)
    post:
      summary: Create a new task
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get task details
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '204':
          description: Task deleted

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'